@@ -1,13 +1,15 @@
-apply plugin: 'com.android.application'
+plugins {
+    id 'com.android.application'
+}
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
+    namespace 'com.example.gpy.whiteboard'
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.example.gpy.whiteboard"
-        minSdkVersion 14
-        targetSdkVersion 23
+        minSdk 24
+        targetSdk 34
         versionCode 1
         versionName "1.0"
     }
@@ -17,12 +19,15 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile project(':wblib')
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile'com.squareup.retrofit:converter-jackson:1.9.0'
-    compile 'com.jakewharton:butterknife:6.1.0'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':wblib')
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.squareup.retrofit:converter-jackson:1.9.0'
 }
