@@ -1,7 +1,7 @@
 package com.example.gpy.whiteboard.view.widget.floatingactionmenu;
 
 import android.graphics.Rect;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.MotionEvent;
 import android.view.TouchDelegate;
 import android.view.View;
